Apaguei as páginas sem rotas e separei as rotas de acordo com adm ou cliente .py
