--- app/templates/admin.html
@@ -1,9 +1,0 @@
-<html>
-    <head>
-        <link rel="stylesheet" href=" {{ url_for('static', filename='css/style.css') }} ">
-    </head>
-    <body>
-        <br>Saudação: {{par1}}
-        <br>Administrador: {{par2}}
-    </body>
-</html>
