--- app/templates/usuario.html
@@ -1,30 +1,0 @@
-<html>
-    <head>
-        <link rel="stylesheet" href=" {{ url_for('static', filename='css/style.css') }} ">
-    </head>
-    <body>
-        <div style="text-align:center">
-            <p>Página do Usuário</p>
-            <table>
-                <tr>
-                    <td><label name="name">Nome:</label></td>
-                    <td>{{name}}</td>
-                </tr>
-                <tr>
-                    <td><label name="cpf">CPF:</label></td>
-                    <td>{{cpf}}</td>
-                </tr>
-                <tr>
-                    <td><label name="user">Usuário:</label></td>
-                    <td>{{user}}</td>
-                </tr>
-                <tr>
-                    <td><label name="password">Senha:</label></td>
-                    <td>{{password}}</td>
-                </tr>
-            </table>
-
-        </div>
-        <br><a href="index">Página Inicial</a>
-    </body>
-</html>
